Fix math utils position
