+++ web/client/utils/styleparser/StyleParserUtils.js
@@ -38,7 +38,7 @@ import axios from 'axios';
 import isNil from 'lodash/isNil';
 import isObject from 'lodash/isObject';
 import MarkerUtils from '../MarkerUtils';
-import {randomInt} from '../cesium/MathUtils';
+import {randomInt} from '../MathUtils';
 
 
 export const isGeoStylerBooleanFunction = (got) => [
